requirements_worker.txt
requests==2.31.0
numpy==1.26.4
redis>=5.3.0,<6.0
huggingface_hub>=0.26.0
psycopg2-binary==2.9.9
pandas==2.3.3
